[docker] moved environment variables to app.config
