--- client/utils/config.js
@@ -1,5 +1,0 @@
-const BASE_URL = 'http://stacking:3000';
-
-module.exports = {
-  BASE_URL,
-};
