+++ docker-compose.yml
@@ -1,7 +1,7 @@
 version: "2"
 services:
   stacking:
-    image: pazano/stacking:v0
+    image: pazano/stacking:v1
     restart: on-failure
     working_dir: /usr/src/app
     ports:
