+++ server/database/index.js
@@ -1,11 +1,12 @@
 const mongoose = require('mongoose');
+const { MONGO_URI } = require('../../app.config');
 
-mongoose.connect('mongodb://mongo/matryoksha');
+mongoose.connect(`${MONGO_URI}`);
 const db = mongoose.connection;
 
 db.on('error', (err) => {
   console.log('Error connecting to database.  Retrying in 3 seconds');
-  setTimeout(() => mongoose.connect('mongodb://mongo/matryoksha'), 3000);
+  setTimeout(() => mongoose.connect(`${MONGO_URI}`), 3000);
 });
 
 db.once('open', () => {
